+++ packages/plugin-unsupported-blocks/web/src/tiptap/tiptap.ts
@@ -1,0 +1,13 @@
+import { CreateRicosExtensions } from 'wix-tiptap-editor';
+import { Component } from '../unsupported-blocks-component';
+
+export const createRicosExtensions: CreateRicosExtensions = defaultOptions => [
+  {
+    type: 'node' as const,
+    Component,
+    createExtensionConfig: name => ({
+      name,
+      defaultOptions,
+    }),
+  },
+];