+++ packages/ricos-content/web/lib/duplicate.ts
@@ -1,0 +1,1 @@
+export { duplicate } from '../src/RicosContentAPI/metadata-utils';