+++ packages/ricos-content/web/lib/metadata-utils.ts
@@ -1,0 +1,1 @@
+export { extractIdFromContent, ensureContentId } from '../src/RicosContentAPI/metadata-utils';