+++ packages/ricos-content/web/src/converters/draft/toDraft/convertStructToJson.ts
@@ -1,46 +1,35 @@
-import { isObject } from 'lodash';
-import { Struct, Value, NullValue, ListValue } from 'ricos-schema';
+import * as A from 'fp-ts/Array';
+import { identity, pipe } from 'fp-ts/function';
+import * as J from 'fp-ts/Json';
+import { concatAll } from 'fp-ts/Monoid';
+import * as R from 'fp-ts/Record';
+import { first } from 'fp-ts/Semigroup';
+import * as T from 'fp-ts/Tuple';
+import { ListValue, Struct, Value } from 'ricos-schema';
+import { firstRight } from '../../../fp-utils';
+import * as SRT from '../struct-refined-types';
 
-export function convertStructToJson(message: Struct): Record<string, unknown> {
-  const json: Record<string, unknown> = {};
-  Object.entries(message.fields).forEach(([k, v]) => {
-    json[k] = convertValueToJson(v);
-  });
-  return json;
-}
+const mergeJsonMap = R.getMonoid(first<J.Json>());
+const toRecord = ([k, v]: [string, J.Json]): J.JsonRecord => ({ [k]: v });
 
-function convertValueToJson(message: Value) {
-  if (
-    !message.boolValue &&
-    !message.nullValue &&
-    !message.numberValue &&
-    !message.stringValue &&
-    !message.listValue &&
-    !message.structValue
-  ) {
-    throw new TypeError('Invalid value');
-  }
-  if (message.boolValue) {
-    return message.boolValue;
-  } else if (message.stringValue) {
-    return message.stringValue;
-  } else if (message.numberValue) {
-    return message.numberValue;
-  } else if (message.nullValue === NullValue.NULL_VALUE) {
-    return null;
-  } else if (message.listValue) {
-    if (!Array.isArray(message.listValue?.values)) {
-      throw new TypeError('Invalid list value');
-    }
-    return convertListToJson(message.listValue);
-  } else if (message.structValue) {
-    if (!isObject(message.structValue.fields)) {
-      throw new TypeError('Invalid struct value');
-    }
-    return convertStructToJson(message.structValue);
-  }
-}
+export const convertStructToJson = (message: Struct): J.JsonRecord =>
+  pipe(
+    message.fields,
+    Object.entries,
+    A.map(T.bimap(convertValueToJson, identity)),
+    A.map(toRecord),
+    concatAll(mergeJsonMap)
+  );
 
-function convertListToJson(message: ListValue): unknown[] {
-  return message.values.map(v => convertValueToJson(v));
-}
+const convertValueToJson = (message: Value): J.Json =>
+  firstRight(message, null as J.Json, [
+    [SRT.isBoolValue, ({ boolValue }) => boolValue as boolean],
+    [SRT.isNumberValue, ({ numberValue }) => numberValue as number],
+    [SRT.isStringValue, ({ stringValue }) => stringValue as string],
+    [SRT.isNullValue, () => null],
+    [SRT.isArrayValue, m => convertListToJson(m.listValue || { values: [] })],
+    [SRT.isStructValue, m => convertStructToJson(m.structValue || { fields: {} })],
+  ]);
+
+const convertListToJson = (message: ListValue): J.JsonArray =>
+  pipe(message.values, A.map(convertValueToJson));