+++ packages/ricos-editor/web/src/toolbars/utils/biUtils.ts
@@ -1,10 +1,11 @@
 import { Version, Helpers, ToolbarType } from 'wix-rich-content-common';
 
-export const getBiFunctions = (helpers: Helpers) => {
+export const getBiFunctions = (helpers: Helpers, contentId?: string) => {
   const onInlineToolbarOpen = toolbarType =>
     helpers?.onInlineToolbarOpen?.({
       toolbarType,
       version: Version.currentVersion,
+      contentId,
     });
   const onToolbarButtonClick = (
     name: string,
@@ -18,10 +19,17 @@ export const getBiFunctions = (helpers: Helpers) => {
       type: toolbarType,
       value: value === undefined ? undefined : typeof value === 'boolean' ? `${!value}` : value,
       version: Version.currentVersion,
+      contentId,
     });
     if (pluginId && value && typeof value === 'string') {
-      helpers?.onPluginAdd?.(pluginId, 'FormattingToolbar', Version.currentVersion);
-      helpers?.onPluginAddSuccess?.(pluginId, 'FormattingToolbar', value, Version.currentVersion);
+      helpers?.onPluginAdd?.(pluginId, 'FormattingToolbar', Version.currentVersion, contentId);
+      helpers?.onPluginAddSuccess?.(
+        pluginId,
+        'FormattingToolbar',
+        value,
+        Version.currentVersion,
+        contentId
+      );
     }
   };
   return { onInlineToolbarOpen, onToolbarButtonClick };