+++ packages/tiptap-editor/web/src/components/RicosTiptapEditor/RicosTiptapEditor.tsx
@@ -10,6 +10,19 @@ import { tiptapExtensions as coreExtensions } from '../../tiptap-extensions';
 import { RicosTiptapEditorProps } from '../../types';
 import { coreConfigs } from './core-configs';
 import { getLangDir } from 'wix-rich-content-common';
+import { Node } from 'prosemirror-model';
+
+// TODO: maybe should move it to utils ?
+const getSelectedNodes = ({ editor }) => {
+  const selection = editor.state.selection;
+  const nodes: Node[] = [];
+  editor.state.doc.nodesBetween(selection.from, selection.to, (node: Node) => {
+    // For not including text nodes inside paragraph/heading nodes etc.
+    node.attrs.id && nodes.push(node);
+  });
+
+  return nodes;
+};
 
 export const RicosTiptapEditor: FunctionComponent<RicosTiptapEditorProps> = ({
   content,
@@ -17,6 +30,7 @@ export const RicosTiptapEditor: FunctionComponent<RicosTiptapEditorProps> = ({
   onLoad,
   onUpdate,
   onSelectionUpdate,
+  onBlur,
   theme,
   locale,
   ...context
@@ -25,6 +39,12 @@ export const RicosTiptapEditor: FunctionComponent<RicosTiptapEditorProps> = ({
   const [editor, setEditor] = useState<Editor>((null as unknown) as Editor);
   const mergedExtensions = Extensions.of([...coreConfigs, ...extensions]);
 
+  const _onUpdate = editor => {
+    const newContent = editor.getJSON();
+    const convertedContent = tiptapToDraft(newContent as JSONContent);
+    onUpdate?.({ content: convertedContent });
+  };
+
   useEffect(() => {
     const tiptapExtensions = mergedExtensions.getTiptapExtensions();
     const editorInstance = new Editor({
@@ -32,18 +52,18 @@ export const RicosTiptapEditor: FunctionComponent<RicosTiptapEditorProps> = ({
       content,
       injectCSS: true,
       onUpdate: ({ editor }) => {
-        const newContent = editor.getJSON();
-        const convertedContent = tiptapToDraft(newContent as JSONContent);
-        onUpdate?.({ content: convertedContent });
-      },
-      onSelectionUpdate: () => {
-        onSelectionUpdate?.();
+        _onUpdate(editor);
       },
       onBlur: () => {
-        onSelectionUpdate?.();
+        onBlur?.();
       },
     });
 
+    editorInstance.on('selectionUpdate', ({ editor }) => {
+      const selectedNodes = getSelectedNodes({ editor });
+      onSelectionUpdate?.({ selectedNodes });
+      _onUpdate(editor);
+    });
     editorInstance.on('transaction', forceUpdate);
 
     setEditor(editorInstance);