+++ packages/toolbars-new/web/src/Toolbar/buttonsListCreator.tsx
@@ -23,13 +23,15 @@ import {
   translateHeading,
   findOsName,
   getSpacing,
+  documentStyleCssProperties,
+  inlineOverrideStyles,
 } from './buttonsListCreatorConsts';
 import { HEADER_TYPE_MAP } from 'wix-rich-content-plugin-commons';
 import {
   convertRelStringToObject,
   convertRelObjectToString,
 } from 'wix-rich-content-common/libs/linkConverters';
-import { getFontSizeNumber, hasStyleChanges } from './utils';
+import { getFontSizeNumber, hasStyleChanges, getBlockStyle } from './utils';
 import style from './ToolbarButtonNew.scss';
 
 type editorCommands = EditorCommands;
@@ -74,7 +76,7 @@ export const createButtonsList = (
         defaultLineSpacing,
         theme
       );
-      handleButtonOnSave(buttonsList, index, editorCommands, headingsData);
+      handleButtonOnSave(buttonsList, index, editorCommands);
       handleButtonOnCancel(buttonsList, index, editorCommands);
       handleButtonOnChange(buttonsList, index, editorCommands, headingsData);
       handleButtonOnDone(buttonsList, index, editorCommands);
@@ -121,7 +123,10 @@ const handleButtonColorPicker = (
 ) => {
   if (buttonsFullData[buttonsList[index].name].type === 'color-picker') {
     const buttonName = buttonsList[index].name;
-    buttonsList[index].getCurrentColor = () => editorCommands.getColor(colorTypes[buttonName]);
+    const blockStyle = getBlockStyle(editorCommands);
+    buttonsList[index].getCurrentColor = () =>
+      editorCommands.getColor(colorTypes[buttonName]) ||
+      blockStyle?.[documentStyleCssProperties[buttonName]];
     buttonsList[index].onColorAdded = color => colorPickerData[buttonName]?.onColorAdded?.(color);
     buttonsList[index].onChange = color => {
       editorCommands.insertDecoration(colorTypes[buttonName], { color });
@@ -239,17 +244,21 @@ const handleTitleButton = (buttonsList, index, editorCommands: editorCommands) =
   buttonsList[index].isActive = () => headingKey === 'header-three' || headingKey === 'header-two';
 };
 
-const handleButtonOnSave = (buttonsList, index, editorCommands: editorCommands, headingsData) => {
+const handleButtonOnSave = (buttonsList, index, editorCommands: editorCommands) => {
   if (buttonsFullData[buttonsList[index].name].onSave) {
     const buttonName = buttonsList[index].name;
     if (Object.keys(textBlockButtons).includes(buttonName)) {
       buttonsList[index].onSave = type => {
-        buttonName === 'HEADINGS' &&
-          headingsData.allowHeadingCustomization &&
+        let shouldSetBlockType = true;
+        if (buttonName === 'HEADINGS') {
           editorCommands.clearSelectedBlocksInlineStyles();
-        setTimeout(() => {
-          editorCommands.setBlockType(type);
-        });
+          const currentHeading = HEADER_TYPE_MAP[getCurrentHeading(editorCommands)];
+          shouldSetBlockType = currentHeading !== type;
+        }
+        shouldSetBlockType &&
+          setTimeout(() => {
+            editorCommands.setBlockType(type);
+          });
       };
     } else if (buttonName === 'Alignment') {
       buttonsList[index].onSave = type => editorCommands.setTextAlignment(type);
@@ -354,8 +363,16 @@ const handleButtonIsDisabled = (buttonsList, index, editorCommands: editorComman
 const handleButtonIsActive = (buttonsList, index, editorCommands: editorCommands) => {
   const buttonName = buttonsList[index].name;
   if (Object.keys(inlineStyleButtons).includes(buttonName)) {
-    buttonsList[index].isActive = () =>
-      editorCommands.hasInlineStyle(inlineStyleButtons[buttonName]);
+    buttonsList[index].isActive = () => {
+      const blockStyle = getBlockStyle(editorCommands);
+      const property = documentStyleCssProperties[buttonName];
+      return (
+        editorCommands.hasInlineStyle(inlineStyleButtons[buttonName]) ||
+        (blockStyle?.[property] &&
+          blockStyle[property] === inlineStyleButtons[buttonName] &&
+          !editorCommands.hasInlineStyle(inlineOverrideStyles[buttonName]))
+      );
+    };
   } else if (Object.keys(textBlockButtons).includes(buttonName)) {
     buttonsList[index].isActive = () =>
       editorCommands.isBlockTypeSelected(textBlockButtons[buttonName]);